Refactored font-size based list measurement logic into a custom hook
